+++ test/alternator/test_key_conditions.py
@@ -9,7 +9,7 @@
 
 import pytest
 from botocore.exceptions import ClientError
-from util import random_string, random_bytes, full_query, multiset
+from alternator_util import random_string, random_bytes, full_query, multiset
 
 # The test_table_{sn,ss,sb}_with_sorted_partition fixtures are the regular
 # test_table_{sn,ss,sb} fixture with a partition inserted with many items.
