+++ test/alternator/test_lsi.py
@@ -11,7 +11,7 @@
 import pytest
 import time
 from botocore.exceptions import ClientError
-from util import create_test_table, new_test_table, random_string, full_scan, full_query, multiset, list_tables
+from alternator_util import create_test_table, new_test_table, random_string, full_scan, full_query, multiset, list_tables
 
 # LSIs support strongly-consistent reads, so the following functions do not
 # need to retry like we did in test_gsi.py for GSIs:
