+++ test/alternator/test_nested.py
@@ -6,7 +6,7 @@
 
 import pytest
 from botocore.exceptions import ClientError
-from util import random_string
+from alternator_util import random_string
 
 # Test that we can write a top-level attribute that is a nested document, and
 # read it back correctly.
