+++ test/alternator/test_query_filter.py
@@ -9,7 +9,7 @@
 import pytest
 from botocore.exceptions import ClientError
 import random
-from util import full_query, full_query_and_counts, random_string, random_bytes
+from alternator_util import full_query, full_query_and_counts, random_string, random_bytes
 
 # The test_table_sn_with_data fixture is the regular test_table_sn fixture
 # with a partition inserted with 20 items. The sort key 'c' of the items
