+++ test/alternator/test_tracing.py
@@ -19,7 +19,7 @@ import requests
 import time
 import json
 from botocore.exceptions import ClientError
-from util import random_string, full_scan, full_query, create_test_table
+from alternator_util import random_string, full_scan, full_query, create_test_table
 
 # The "with_tracing" fixture ensures that tracing is enabled throughout
 # the run of a test function, and disabled when it ends. If tracing cannot be
